Create Lecture 1 problem set - Credit
The advanced option for the CS50 problem sets. 

The requirements for this project were to:
receive a long number for the credit card, and reject any numbers that don't fit the criteria, so numbers with hyphens or negative numbers for example. 
Than, to use and implement Luhn’s Algorithm into the code, to verify that the credit card number provided was valid.
and finally, based on the length of the number and other factors, to determine which credit card company it belonged to, for example "VISA" or "MASTERCARD"